Align pros/cons with the hovered option
Highlight the hovered popup option.
Tweak pro/con styling.
Use padding above/below controls instead of margin-bottom, so that we can set the background and have the label centered.
Set the id on control groups.
Close the popup if its inner and outer width are resized to be the same.